backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
msal==1.24.1
openai==1.54.4
psycopg2-binary==2.9.10
gunicorn==21.2.0
email-validator==2.1.0
Werkzeug==3.1.3